Block color(buggy), block value, timelines,
Block values are between 1 and 25. Blocks are color coded.
Seperate timelines for block, magnet, shield. Seperate power timelines for shield, magnet.

+++ Snakey/src/Coin.java
@@ -105,7 +105,7 @@ public class Coin implements Serializable{
         c.setTranslateY(imageView.getY()+10);
         root.getChildren().addAll(imageView,c);
     }
-    public void updatemovement(Pane root, int speed) {
+    public void updatemovement(Pane root, double speed) {
     	if(imageView!=null) 
     	{
 			if(imageView.getY()<640) {
